# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Prime Link OS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  status: string;
  online: boolean;
  /** @nullable */
  lastActive?: string | null;
  /** @nullable */
  performanceScore?: number | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface CreateUserBody {
  name: string;
  email: string;
  password: string;
  role: string;
}

export interface UpdateUserBody {
  name?: string;
  role?: string;
  status?: string;
  /** @nullable */
  performanceScore?: number | null;
}

export interface UpdateStatusBody {
  online: boolean;
}

export interface Client {
  id: number;
  clientName: string;
  phone: string;
  /** @nullable */
  business?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  planId?: number | null;
  /** @nullable */
  addedBy?: number | null;
  /** @nullable */
  assignedTo?: number | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CreateClientBody {
  clientName: string;
  phone: string;
  business?: string;
  website?: string;
  /** @nullable */
  planId?: number | null;
  notes?: string;
}

export interface UpdateClientBody {
  clientName?: string;
  phone?: string;
  business?: string;
  website?: string;
  /** @nullable */
  planId?: number | null;
  /** @nullable */
  assignedTo?: number | null;
  status?: string;
  notes?: string;
}

export interface Task {
  id: number;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  assignedTo?: number | null;
  title: string;
  /** @nullable */
  description?: string | null;
  priority: string;
  status: string;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  resultUrl?: string | null;
  createdAt: string;
}

export interface CreateTaskBody {
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  assignedTo?: number | null;
  title: string;
  description?: string;
  priority: string;
  /** @nullable */
  deadline?: string | null;
}

export interface UpdateTaskBody {
  title?: string;
  description?: string;
  priority?: string;
  status?: string;
  /** @nullable */
  assignedTo?: number | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  progress?: number | null;
  resultUrl?: string;
}

export interface Earning {
  id: number;
  userId: number;
  amount: number;
  type: string;
  status: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  taskId?: number | null;
  /** @nullable */
  clientId?: number | null;
  createdAt: string;
}

export interface CreateEarningBody {
  userId: number;
  amount: number;
  type: string;
  description?: string;
  /** @nullable */
  taskId?: number | null;
  /** @nullable */
  clientId?: number | null;
}

export interface UpdateEarningBody {
  status?: string;
  amount?: number;
}

export interface Message {
  id: number;
  senderId: number;
  /** @nullable */
  receiverId?: number | null;
  isGroup: boolean;
  content: string;
  /** @nullable */
  fileUrl?: string | null;
  isPinned: boolean;
  isRead: boolean;
  createdAt: string;
}

export interface CreateMessageBody {
  /** @nullable */
  receiverId?: number | null;
  isGroup: boolean;
  content: string;
  fileUrl?: string;
}

export interface Plan {
  id: number;
  name: string;
  clientPrice: number;
  salesmanCommission: number;
  workerPayment: number;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface CreatePlanBody {
  name: string;
  clientPrice: number;
  salesmanCommission: number;
  workerPayment: number;
  description?: string;
}

export interface UpdatePlanBody {
  name?: string;
  clientPrice?: number;
  salesmanCommission?: number;
  workerPayment?: number;
  description?: string;
}

export interface DashboardStats {
  totalRevenue: number;
  revenueThisMonth: number;
  totalEmployees: number;
  onlineEmployees: number;
  ongoingTasks: number;
  completedTasks: number;
  pendingApprovals: number;
  totalClients: number;
  newClientsThisMonth: number;
  totalSalaryOwed: number;
  profit: number;
}

export interface RevenueByMonth {
  month: string;
  revenue: number;
  salary: number;
  profit: number;
}

export interface PerformerStat {
  userId: number;
  name: string;
  count: number;
  totalEarnings: number;
}

export interface TopPerformers {
  topSalesmen: PerformerStat[];
  topWorkers: PerformerStat[];
}

export type ListClientsParams = {
  status?: string;
  /**
   * @nullable
   */
  addedBy?: number | null;
};

export type ListTasksParams = {
  /**
   * @nullable
   */
  assignedTo?: number | null;
  status?: string;
  priority?: string;
};

export type ListEarningsParams = {
  /**
   * @nullable
   */
  userId?: number | null;
  status?: string;
};

export type ListMessagesParams = {
  /**
   * @nullable
   */
  receiverId?: number | null;
  isGroup?: boolean;
};